+++ SmartStudent/src/main/java/com/manage/admin/service/Admin_StudentSemesterServiceImpl.java
@@ -7,6 +7,8 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
+
 @Service
 @Qualifier("adminStudentSemesterServiceImpl")
 public class Admin_StudentSemesterServiceImpl implements Admin_StudentSemesterService {
@@ -21,7 +23,9 @@ public class Admin_StudentSemesterServiceImpl implements Admin_StudentSemesterSe
     @Override
     @Transactional(transactionManager = "adminTransactionManager")
     public StudentSemester getStudentSemesterById(Long studentSemesterId) {
-        return adminStudentSemesterDAO.getStudentSemesterById(studentSemesterId);
+        StudentSemester studentSemester = adminStudentSemesterDAO.getStudentSemesterById(studentSemesterId);
+        initializeStudentSemester(studentSemester);
+        return studentSemester;
     }
 
     @Override
@@ -40,6 +44,29 @@ public class Admin_StudentSemesterServiceImpl implements Admin_StudentSemesterSe
     @Transactional(transactionManager = "adminTransactionManager")
     public void deleteStudentSemester(Long studentSemesterId) {
         StudentSemester studentSemester = getStudentSemesterById(studentSemesterId);
-        adminStudentSemesterDAO.deleteStudentSemester(studentSemester);
+        if (studentSemester != null) {
+            adminStudentSemesterDAO.deleteStudentSemester(studentSemester);
+        }
+    }
+
+    @Override
+    @Transactional(transactionManager = "adminTransactionManager")
+    public List<StudentSemester> getAllStudentSemesters() {
+        List<StudentSemester> studentSemesters = adminStudentSemesterDAO.getAllStudentSemesters();
+        studentSemesters.forEach(this::initializeStudentSemester);
+        return studentSemesters;
+    }
+    
+    
+
+    private void initializeStudentSemester(StudentSemester studentSemester) {
+        if (studentSemester != null) {
+            if (studentSemester.getStudent() != null) {
+                studentSemester.getStudent().getUsername(); // Force initialization of the Student entity
+            }
+            if (studentSemester.getSemester() != null) {
+                studentSemester.getSemester().getName(); // Force initialization of the Semester entity
+            }
+        }
     }
 }
